+++ TrainingLoop/TrainingLoop.swift
@@ -18,7 +18,7 @@ import TensorFlow
 // Workaround https://bugs.swift.org/browse/TF-1122 that prevents us from registering a
 // loss function inside our TrainingLoop struct
 public final class LossFunctionWrapper<Output: Differentiable, Target> {
-  public typealias F = @differentiable (Output, @noDerivative Target) -> Tensor<Float>
+  public typealias F = @differentiable(Output, @noDerivative Target) -> Tensor<Float>
   public var f: F
   init(_ f: @escaping F) { self.f = f }
 }
@@ -34,25 +34,32 @@ public protocol TrainingLoopProtocol {
   where
     Training: Sequence, Training.Element: Collection,
     Training.Element.Element == LabeledData<Opt.Model.Input, Target>
+
   /// The type of the collection of batches for the validation data.
   associatedtype Validation
   where
     Validation: Collection,
     Validation.Element == LabeledData<Opt.Model.Input, Target>
+
   /// The type of the target of our model.
   associatedtype Target
+
   /// The type of the optimizer used.
   associatedtype Opt: Optimizer where Opt.Model: Module
 
   // Typealiases
   /// The type of the model.
   typealias Model = Opt.Model
+
   /// The type of the input of the model.
   typealias Input = Opt.Model.Input
+
   /// The type of the output of the model.
   typealias Output = Opt.Model.Output
+
   /// The type of a batch.
   typealias Batch = LabeledData<Input, Target>
+
   // In a wrapper for now because of TF-1122.
   /// The type of the loss function.
   typealias LossFunction = LossFunctionWrapper<Output, Target>
@@ -60,16 +67,19 @@ public protocol TrainingLoopProtocol {
   // Data
   /// The training epochs.
   var training: Training { get }
+
   /// The validation batches.
   var validation: Validation { get }
 
   // Optimizer and loss function
   /// The optimizer.
   var optimizer: Opt { get set }
+
   /// The loss function.
   var lossFunction: LossFunction { get set }
+
   /// The metrics
-  var metrics: [TrainingMetrics] { get set}
+  var metrics: [TrainingMetrics] { get set }
 
   // Callbacks
   /// The callbacks used to customize the training loop.
@@ -81,16 +91,22 @@ public protocol TrainingLoopProtocol {
 
   /// The last input fed to the model.
   var lastStepInput: Input? { get set }
+
   /// The last target.
   var lastStepTarget: Target? { get set }
+
   /// The last predictions of the model.
   var lastStepOutput: Output? { get set }
+
   /// The last gradients computed.
   var lastStepGradient: Model.TangentVector? { get set }
+
   /// The last loss.
   var lastStepLoss: Tensor<Float>? { get set }
+
   /// The number of batches in the current collection of batches.
   var batchCount: Int? { get set }
+
   /// The index of the current batch.
   var batchIndex: Int? { get set }
 
@@ -98,9 +114,10 @@ public protocol TrainingLoopProtocol {
 
   /// The number of epochs we are currently fitting for.
   var epochCount: Int? { get set }
+
   /// The index of the current epoch.
   var epochIndex: Int? { get set }
-  
+
   // MARK: - Others
 
   /// The log for last statistics
@@ -114,26 +131,37 @@ public protocol TrainingLoopProtocol {
 public enum TrainingLoopEvent {
   /// The start of a fit.
   case fitStart
+
   /// The end of a fit.
   case fitEnd
+
   /// The start of one epoch (training + validation).
   case epochStart
+
   /// The start of one epoch (training + validation).
   case epochEnd
+
   /// The start of a training phase.
   case trainingStart
+
   /// The end of a training phase.
   case trainingEnd
+
   /// The start of a validation phase.
   case validationStart
+
   /// The end of a validation phase.
   case validationEnd
+
   /// The start of a training or inference step on a batch.
   case batchStart
+
   /// The end of a training or inference step on a batch.
   case batchEnd
+
   /// At the start of the optimizer update, just after the differentiable step.
   case updateStart
+
   /// Just after the model prediction at inference, before computing the loss.
   case inferencePredictionEnd
 }
@@ -159,12 +187,16 @@ where
   // Typealiases
   /// The type of the model.
   public typealias Model = Opt.Model
+
   /// The type of the input of the model.
   public typealias Input = Opt.Model.Input
+
   /// The type of the output of the model.
   public typealias Output = Opt.Model.Output
+
   /// The type of a batch.
   public typealias Batch = LabeledData<Input, Target>
+
   // In a wrapper for now because of TF-1122.
   /// The type of the loss function.
   public typealias LossFunction = LossFunctionWrapper<Output, Target>
@@ -172,14 +204,17 @@ where
   // Data
   /// The training epochs.
   public let training: Training
+
   /// The validation batches.
   public let validation: Validation
 
   // Optimizer and loss function
   /// The optimizer.
   public var optimizer: Opt
+
   /// The loss function
   public var lossFunction: LossFunction
+
   /// The metrics
   public var metrics: [TrainingMetrics]
 
@@ -193,23 +228,30 @@ where
 
   /// The last input fed to the model.
   public var lastStepInput: Input? = nil
+
   /// The last target.
   public var lastStepTarget: Target? = nil
+
   /// The last predictions of the model.
   public var lastStepOutput: Output? = nil
+
   /// The last gradients computed.
   public var lastStepGradient: Model.TangentVector? = nil
+
   /// The last loss.
   public var lastStepLoss: Tensor<Float>? = nil
+
   /// The number of batches in the current collection of batches.
   public var batchCount: Int? = nil
+
   /// The index of the current batch.
   public var batchIndex: Int? = nil
 
   // MARK: - Epoch-level data
-  
+
   /// The number of epochs we are currently fitting for.
   public var epochCount: Int? = nil
+
   /// The index of the current epoch.
   public var epochIndex: Int? = nil
 
@@ -220,6 +262,7 @@ where
 
   /// Default callback objects
   public let statisticsRecorder: StatisticsRecorder
+
   public let progressPrinter: ProgressPrinter
 
   /// Creates an instance from `training` and `validation` data, a `model`, an `optimizer` and a
@@ -228,7 +271,7 @@ where
   /// Parameter callbacks: Callbacks that the `TrainingLoop` will use in every call to fit.
   public init(
     training: Training, validation: Validation, optimizer: Opt,
-    lossFunction: @escaping LossFunction.F, 
+    lossFunction: @escaping LossFunction.F,
     metrics: [TrainingMetrics] = [],
     callbacks: [TrainingLoopCallback<Self>] = []
   ) {
@@ -242,8 +285,8 @@ where
     self.statisticsRecorder = statisticsRecorder
     self.progressPrinter = progressPrinter
     self.callbacks = [
-      statisticsRecorder.record, 
-      progressPrinter.print
+      statisticsRecorder.record,
+      progressPrinter.print,
     ] + callbacks
   }
 }
@@ -253,7 +296,8 @@ extension TrainingLoop {
   public mutating func differentiableStep(model: Model) throws {
     guard let data = lastStepInput else { return }
     guard let target = lastStepTarget else { return }
-    (lastStepLoss, lastStepGradient) = valueWithGradient(at: model) { (model: Model) -> Tensor<Float> in
+    (lastStepLoss, lastStepGradient) = valueWithGradient(at: model) {
+      (model: Model) -> Tensor<Float> in
       let predictions = model(data)
       lastStepOutput = predictions
       return lossFunction.f(predictions, target)
@@ -285,12 +329,16 @@ extension TrainingLoop {
 public enum TrainingLoopAction: Error {
   /// Abort actions in the current training/inference step and goes to the next batch.
   case cancelBatch
+
   /// Abort actions in the current training phase and goes to the validation phase.
   case cancelTraining
+
   /// Abort actions in the current validation phase and goes to the next epoch.
   case cancelValidation
+
   /// Abort actions in the current epoch and goes to the next epoch.
   case cancelEpoch
+
   /// Abort actions in the current fit and ends fitting.
   case cancelFit
 }
@@ -334,7 +382,9 @@ extension TrainingLoop {
   public mutating func fit(
     _ model: inout Model, epochs: Int, callbacks: [TrainingLoopCallback<Self>] = [],
     on device: Device = Device.default,
-    differentiableStep: (Model, inout Self) throws -> Void = { try $1.differentiableStep(model: $0) }
+    differentiableStep: (Model, inout Self) throws -> Void = {
+      try $1.differentiableStep(model: $0)
+    }
   ) throws {
     let callbacksCount = self.callbacks.count
     self.callbacks += callbacks
